working with the log file is improved. (#45)

+++ internal/config/log.go
@@ -5,7 +5,7 @@ type Log struct {
 	// directory where log files are stored
 	DirLog string `yaml:"dirLog"`
 	// maximum size of a log file, in megabytes
-	FileSizeMB uint64 `yaml:"fileSizeInMb"`
+	FileSizeMB float64 `yaml:"fileSizeInMb"`
 	// number of records to show when
 	APIShowRecords uint64 `yaml:"apiShowRecords"`
 }
